requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-filter==2.3.0
django-jalali==4.0.0
django-render-partial==0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
httpie==2.2.0
idna==2.10
inflection==0.5.0
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
mysqlclient==2.0.1
django-cors-headers==3.4.0

